javascript:chore - use a better approach to test parsing
Previously the tests of parsing Javascript to a generics AST was a bit
tedious and redundant. Basically, a source code was used as input and
the entire expected AST needed to be declared, which make the process of
adding new test cases tedious and redundant, since this AST contains a
lot of nested structs.

This commit change the tests workflow to use a source code file as input
and a file with the contents of the AST to use as expected. This
expected file contains the output from `ast.Print`, and the
`ast.Fprint` is used to compare the expected vs actual ASTs. Now, to add
new test cases its only necessary create a new input file on
`testdata/source` and the expected output AST on `testdata/expected`.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>