ir:feat - add support for file imports
This commit add a support to handle file imports on IR.

To support file imports a new interface Member was created. This
interface basically represents a object member that can be created or
declared inside a file; for now this Member could be a Function or a
Import, but in the future this member could be class declarations,
constants and global variable declarations.

A new struct ExternalMember was created that implements Member
interface. This struct is a generic member that represents imports, your
fields holds information about the path of import, the name of import
and a optional alias. These informations is used to construct a concrete
Member like a function or a variable.

A new struct File was also created that holds all members declared on
some file. This struct represents an ast.File converted to IR
representation.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>