ir:chore - create a new test infrastructure
This commit introduce a generic test mechanism to test CST to AST and
AST to IR conversion.

This mechanism is an extension introduced by 0fa4c35, which use this
logic to any CST to AST conversion and AST to IR conversion. The IR
source code tests is the same used by AST testing, so we can ensure that
to a given code we successfully parse to AST and them to IR.

Note: Since the AST to IR conversion is not fully implemented the
TestParseIR tests is falling.

Read the header of internal/semantic_test.go for more details on these tests.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>