ir:feat - add initial support for ast.SelectorExpr
This commit implement a initial support to convert ast.SelectorExpr to
IR representation.

This initial support only handle cases when the function expression of a
call expression is an ast.SelectorExpr e.g `console.log("foo")`. Also
this initial implementation is not recursive, which means the cases like
`a.b.c()` or `a.b().c` will panic.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>