ir:feat: add support for labeled statements
This commit add support for labeled statements (for/while/switch).

To add this support a new struct `lblock` was created to store the
blocks that is used to jump when parsing for/while/switch statements
body. A stack is created to store these blocks when processing the
statements, basically, when entering a for/while/switch stmt we push a
new target and after the processing the stmt body we pop the target and
we processing an `break` or `continue` we just get the head of stack to
jump it. A map of labeled blocks was also implemented to store blocks
that are labeled, so if a block has a label we get the block to jump
from this map, otherwise we get the head of stack to jump it.

This commit also improve the generated AST for labeled statements,
instead of store a list of []ast.Stmt we just store the body of labeled
block, which makes parse in the IR easier.

Signed-off-by: Matheus Alcantara <matheus.alcantara@zup.com.br>